chore: Reorganize and add new folder handlers.
